First iOS commit.
Added the general layout on the storyboard, but need input on how we
want every particular screen to work. Implemented a very basic login
screen. If you type “host” into the username field, the view will segue
to the host side of the app. Likewise if you type “attendee” into the
username field the view will segue to the attendee side of the app. You
can logout of either section using by going to profile and pressing
logout.